Backend/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.4
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
idna==3.4
openpyxl==3.0.10
PyJWT==2.6.0
python-dotenv==0.21.0
pytz==2022.7
requests==2.28.1
sqlparse==0.4.3
twilio==7.16.0
urllib3==1.26.13
whitenoise==6.2.0
